refactor new code between client and firstload
