+++ sdk/highlight-run/src/client/types/Hooks.ts
@@ -1,4 +1,4 @@
-import { LDContext } from 'types/LDContext'
+import { LDContext } from './LDContext'
 
 /**
  * Contextual information provided to evaluation stages.
